修改

+++ src/main/java/com/example/tmapi/entity/ContrastSaleData.java
@@ -3,13 +3,14 @@ package com.example.tmapi.entity;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class ContrastSaleData implements Serializable {
 
     private Date date;
-    private float lxamt;
+    private BigDecimal lxamt;
     private Date StartDate;
     private Date endDate;
 }
