+++ src/main/java/com/example/tmapi/entity/ContrastStoreSaleData.java
@@ -3,11 +3,12 @@ package com.example.tmapi.entity;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class ContrastStoreSaleData implements Serializable {
     private Date date;
     private String storeId;
-    private float lxamt;
+    private BigDecimal lxamt;
 }
